+++ changelog.md
@@ -5,6 +5,7 @@
 - Add some fake JEI recipes for filling buckets in the bottling machine (BluSunrize)
 - Add a new API interface that allows pressurized fluid in- and output above normal levels (BluSunrize)
 - Add fermentation recipe to turn beetroots into ethanol (BluSunrize)
+- Add arc smelting recipe for raw metal blocks (BluSunrize)
 - Change Electrician villager workstation to be the Circuit Table (BluSunrize)
 - Update JEI handlers for Coke Oven, Blast Furnace and Alloy Smelter to have some animated doodads (BluSunrize)
 - Update manual entry for fluid pipes to document that they can be dyed (BluSunrize)